move "parse" from seTestUtils to parser
