+++ sonarts-core/src/tools/cfg_viewer/viewer.ts
@@ -30,7 +30,7 @@ class Viewer {
   }
 
   public show(source: string) {
-    const sourceFile = parseString(source);
+    const { sourceFile } = parseString(source);
     const graph = ControlFlowGraph.fromStatements(Array.from(sourceFile.statements));
 
     if (!graph) {
